refactored integration script code structure
