+++ ts-client/tests/integration/constants.ts
@@ -1,0 +1,1 @@
+export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
