+++ ts-client/tests/integration/events.ts
@@ -1,0 +1,47 @@
+// TODO: add source from where it copied from CLI.
+import assert from "node:assert";
+import { LogDescription, TransactionReceipt } from "ethers";
+
+// Log Parsing Module.
+type GetEventValueArgs<T extends string, U extends Contract<T>> = {
+  txReceipt: TransactionReceipt;
+  contract: U;
+  eventName: T;
+  value: string;
+};
+
+type Contract<T> = {
+  getEvent(name: T): {
+    fragment: { topicHash: string };
+  };
+  interface: {
+    parseLog(log: { topics: string[]; data: string }): LogDescription | null;
+  };
+};
+
+export function getEventValue<T extends string, U extends Contract<T>>({
+  txReceipt,
+  contract,
+  eventName,
+  value,
+}: GetEventValueArgs<T, U>) {
+  const { topicHash } = contract.getEvent(eventName).fragment;
+
+  const log = txReceipt.logs.find((log) => {
+    return log.topics[0] === topicHash;
+  });
+
+  assert(
+    log !== undefined,
+    `Event '${eventName}' with hash '${topicHash}' not found in logs of the transaction`,
+  );
+
+  const res: unknown = contract.interface
+    .parseLog({
+      data: log.data,
+      topics: [...log.topics],
+    })
+    ?.args.getValue(value);
+
+  return res;
+}
